requirements.txt
streamlit~=1.43.1
google-generativeai
pandas~=2.2.3
python-dotenv~=1.0.1

protobuf~=5.29.3
matplotlib~=3.10.1
plotly~=6.0.0
reportlab~=4.3.1
vaderSentiment~=3.3.2